+++ src/shared_utilities/implicit_timestepping.jl
@@ -102,7 +102,9 @@ function FieldMatrixWithSolver(Y::ClimaCore.Fields.FieldVector)
     explicit_vars = (
         @name(soil.∫F_vol_liq_water_dt),
         @name(soil.∫F_e_dt),
-        @name(soilco2.C),
+        @name(soilco2.CO2),
+        @name(soilco2.O2_f),
+        @name(soilco2.SOC),
         @name(soil.θ_i),
         @name(canopy.hydraulics.ϑ_l),
         @name(snow.S),